+++ packages/use/index.ts
@@ -2,3 +2,4 @@ export * from './src/watch/index';
 export * from './src/whenever/index';
 export * from './src/wheneverEffectScope/index';
 export * from './src/useCountdown/index';
+export * from './src/deepUnref/index';